agent/requirements.txt
strands-agents>=1.26.0
strands-agents-tools>=0.1.0
httpx>=0.28.0
